+++ crowdin_api/api_resources/__init__.py
@@ -1,4 +1,4 @@
-from .applications.resource import ApplicationResource
+from .application.resource import ApplicationResource
 from .bundles.resource import BundlesResource
 from .dictionaries.resource import DictionariesResource
 from .distributions.resource import DistributionsResource
